% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMode.R
\name{getMode}
\alias{getMode}
\title{Get the mode of given values}
\usage{
getMode(values)
}
\arguments{
\item{values}{A numeric vector with at least one element}
}
\value{
The mode of the values
}
\description{
Get the mode of given values
}
\examples{
x <- c(1,2,3,3,3,4,4)
getMode(x)
}
